requirements.txt
streamlit
google-generativeai
edge-tts
SpeechRecognition
streamlit-mic-recorder
pygame